changed from sqlite to postgresql

+++ backend/db/migrate/20210326004504_create_alerts.rb
@@ -6,7 +6,7 @@ class CreateAlerts < ActiveRecord::Migration[6.1]
       t.string :errorCategory
       t.string :errorMessage
       t.string :longMessage
-      t.integer :errorTime
+      t.bigint :errorTime
       t.boolean :selected
       t.boolean :new
       t.boolean :expanded
